+++ backend/module/resolver/response_test.go
@@ -1,0 +1,7 @@
+package resolver
+
+func TestTruncate() {
+	resp := newResponse()
+
+	protojson.
+}
